# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FastAPI学习项目",
    "description": "这是一个用于学习FastAPI框架的示例项目，包含了基础的CRUD操作、用户认证等功能",
    "version": "1.0.0"
  },
  "paths": {
    "/posts/": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "Read Posts",
        "description": "获取文章列表\n\n参数:\n- skip: 跳过的记录数，用于分页\n- limit: 返回的最大记录数，用于分页\n- db: 数据库会话依赖",
        "operationId": "read_posts_posts__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "title": "Response Read Posts Posts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "文章管理"
        ],
        "summary": "Create Post",
        "description": "创建新文章\n\n只有登录用户才能创建文章\n\n参数:\n- post: 文章创建数据\n- db: 数据库会话依赖\n- current_user: 当前登录用户\n\n返回:\n- 创建的文章信息\n\n错误:\n- 401: 未认证\n- 404: 分类不存在",
        "operationId": "create_post_posts__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/published/": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "Read Published Posts",
        "description": "获取已发布的文章列表",
        "operationId": "read_published_posts_posts_published__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "title": "Response Read Published Posts Posts Published  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "Read Post",
        "description": "根据文章ID获取文章信息",
        "operationId": "read_post_posts__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "文章管理"
        ],
        "summary": "Update Post",
        "description": "更新文章\n\n只有文章作者或超级用户才能更新文章\n\n参数:\n- post_id: 文章ID\n- post_update: 文章更新数据\n- db: 数据库会话依赖\n- current_user: 当前登录用户\n\n返回:\n- 更新后的文章信息\n\n错误:\n- 404: 文章不存在\n- 403: 权限不足",
        "operationId": "update_post_posts__post_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "文章管理"
        ],
        "summary": "Delete Post",
        "description": "删除文章\n\n只有文章作者或超级用户才能删除文章\n\n参数:\n- post_id: 文章ID\n- db: 数据库会话依赖\n- current_user: 当前登录用户\n\n返回:\n- 删除成功的消息\n\n错误:\n- 404: 文章不存在\n- 403: 权限不足",
        "operationId": "delete_post_posts__post_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/search/": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "Search Posts",
        "description": "搜索文章",
        "operationId": "search_posts_posts_search__get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  },
                  "title": "Response Search Posts Posts Search  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/": {
      "get": {
        "tags": [
          "分类管理"
        ],
        "summary": "Read Categories",
        "description": "获取分类列表",
        "operationId": "read_categories_categories__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  },
                  "title": "Response Read Categories Categories  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "分类管理"
        ],
        "summary": "Create Category",
        "description": "创建新分类\n\n只有超级用户才能创建分类\n\n参数:\n- category: 分类创建数据\n- db: 数据库会话依赖\n- current_user: 当前超级用户\n\n返回:\n- 创建的分类信息\n\n错误:\n- 400: 分类名称已存在\n- 403: 权限不足",
        "operationId": "create_category_categories__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "Login",
        "description": "用户登录端点\n\n这个端点允许用户使用用户名/邮箱和密码登录\n成功登录后返回JWT访问令牌\n\n参数:\n- user_credentials: 包含用户名/邮箱和密码的登录信息\n- db: 数据库会话依赖\n\n返回:\n- 包含访问令牌、令牌类型、过期时间和用户信息的字典\n\n错误:\n- 401: 用户名/邮箱或密码错误",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "Get Current User Info",
        "description": "获取当前用户信息\n\n这个端点返回当前登录用户的详细信息\n需要有效的JWT令牌\n\n参数:\n- current_user: 当前登录用户（通过JWT令牌验证）\n\n返回:\n- 当前用户的详细信息\n\n错误:\n- 401: 令牌无效或已过期\n- 400: 用户账户已被禁用",
        "operationId": "get_current_user_info_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "Refresh Token",
        "description": "刷新访问令牌\n\n这个端点允许用户使用有效的令牌获取新的令牌\n用于延长用户的登录状态\n\n参数:\n- current_user: 当前登录用户（通过JWT令牌验证）\n\n返回:\n- 新的访问令牌信息\n\n错误:\n- 401: 令牌无效或已过期\n- 400: 用户账户已被禁用",
        "operationId": "refresh_token_auth_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Refresh Token Auth Refresh Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "文章标题"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "文章内容"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "文章摘要"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "description": "是否发布",
            "default": false
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id",
            "description": "分类ID"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "id",
          "author_id",
          "created_at",
          "updated_at"
        ],
        "title": "Post",
        "description": "文章响应模式",
        "examples": [
          {
            "author_id": 1,
            "category_id": 1,
            "content": "用于验证接口的文章内容",
            "created_at": "2025-11-25T09:50:57",
            "id": 8,
            "is_published": true,
            "summary": "接口测试摘要",
            "title": "接口测试文章",
            "updated_at": "2025-11-25T09:50:57"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}